Rename user_defines.h to user_config.h
