+++ sdk/bare/common/drv/analog.h
@@ -1,7 +1,7 @@
 #ifndef ANALOG_H
 #define ANALOG_H
 
-#include "usr/user_defines.h"
+#include "usr/user_config.h"
 #include <stdbool.h>
 #include <stdint.h>
