+++ sdk/bare/common/sys/cmd/cmd_log.c
@@ -1,4 +1,4 @@
-#include "usr/user_defines.h"
+#include "usr/user_config.h"
 
 #if USER_CONFIG_ENABLE_LOGGING == 1
